requirements.txt
# Healthcare Portal - Python Dependencies
# Updated: 2025-10-14

# Core Framework
Flask==3.0.3
Werkzeug==3.0.3

# Authentication & Security
Flask-Login==0.6.3
bcrypt==4.2.0
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.2.0

# Database
supabase==2.7.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Security & Rate Limiting
Flask-CORS==4.0.1
Flask-Limiter==3.8.0
Flask-Session==0.8.0
Flask-Caching==2.3.0

# Utilities
requests==2.32.3
validators==0.33.0
python-dateutil==2.9.0
pytz==2024.1

# PDF Generation
reportlab==4.2.2
weasyprint==62.3

# Email Support
Flask-Mail==0.10.0

# Data Validation & Sanitization
bleach==6.1.0
html5lib==1.1

# Testing (Development)
pytest==8.3.2
pytest-flask==1.3.0
pytest-cov==5.0.0
coverage==7.6.0

# Code Quality (Development)
black==24.8.0
flake8==7.1.1
pylint==3.2.6

# API Documentation
flask-swagger-ui==4.11.1
